Revert "Testing code"
