--- FXApp/.gitignore
@@ -1,1 +1,0 @@
-/MainFXApp.class
