--- View/GoogleMap.fxml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import com.lynden.gmapsfx.GoogleMapView?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controller.GoogleMapController">
-   <children>
-      <GoogleMapView fx:id="mapView" layoutY="15.0" prefHeight="385.0" prefWidth="600.0" />
-      <TextField fx:id="addressField" layoutX="10.0" layoutY="50.0" onAction="#addressTextFieldAction" prefHeight="27.0" prefWidth="274.0" promptText="Address" />
-   </children>
-</AnchorPane>
