--- View/WaterReportView.fxml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controller.WaterReportViewController">
-   <children>
-      <Button layoutX="274.0" layoutY="352.0" mnemonicParsing="false" onAction="#handleCancelPressed" text="Cancel" />
-   </children>
-</AnchorPane>
