+++ View/WelcomeScreen.fxml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuBar?>
+<?import javafx.scene.control.MenuItem?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
@@ -8,14 +11,28 @@
 
 <VBox prefHeight="400.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controller.WelcomeScreenController">
   <children>
+    <MenuBar VBox.vgrow="NEVER">
+      <menus>
+        <Menu mnemonicParsing="false" text="File">
+          <items>
+            <MenuItem mnemonicParsing="false" onAction="#handleCloseMenu" text="Quit" />
+          </items>
+        </Menu>
+        <Menu mnemonicParsing="false" text="Help">
+          <items>
+            <MenuItem mnemonicParsing="false" onAction="#handleAboutMenu" text="About" />
+          </items>
+        </Menu>
+      </menus>
+    </MenuBar>
     <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="-1.0" prefWidth="-1.0" VBox.vgrow="ALWAYS">
       <children>
-            <Text layoutX="178.0" layoutY="184.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Welcome">
+            <Text layoutX="189.0" layoutY="131.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Welcome">
                <font>
                   <Font name="Segoe Print" size="64.0" />
                </font></Text>
-            <Button layoutX="212.0" layoutY="253.0" mnemonicParsing="false" onAction="#loginPressed" text="Login" />
-            <Button layoutX="329.0" layoutY="253.0" mnemonicParsing="false" onAction="#registerPressed" text="Registration" />
+            <Button layoutX="201.0" layoutY="200.0" mnemonicParsing="false" onAction="#loginPressed" text="Login" />
+            <Button layoutX="340.0" layoutY="200.0" mnemonicParsing="false" onAction="#registerPressed" text="Registration" />
       </children>
     </AnchorPane>
   </children>
