--- build.gradle
@@ -1,7 +1,0 @@
-apply plugin: 'java'
-
-dependencies {
-    compile files('lib/GMapsFX-2.0.9.jar')
-    compile files('lib/GMapsFX-2.0.9-sources.jar')
-    compile files('lib/GMapsFX-2.0.9-javadoc.jar')
-}
